+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Improved
 
 * [#55](https://github.com/sdss/lvmgort/pull/55) Better messaging for alerts unavailability. A resolution alert message is now sent when the alerts become available again.
+* Test Python 3.14 in CI.
 
 ### 🔧 Fixed